refactor: (#740) 알림 리스트가 포함된 알림 컨테이너 폴더 이동
- component/common > component
