+++ frontend/src/components/AlarmContainer/ListStyle.ts
