+++ frontend/src/components/AlarmContainer/style.ts
